refactor: merge new patrol code with old, will extract to module later
