+++ modules/ept-posts/client/patrol/patrol.data.html
@@ -66,13 +66,11 @@
       </div>
       <!-- Thread title -->
       <div class="thread-title">
-        <a ui-sref="threads.data({ boardId: post.board_id })">
-          <span ng-bind-html="post.board_name"></span>
-        </a>
-        /
-        <a ui-sref="posts.data({ threadId: post.thread_id, start: post.position, '#': post.id })">
-          <span ng-bind-html="post.thread_title"></span>
-        </a>
+        <span ng-repeat="breadcrumb in post.breadcrumbs" ng-if="breadcrumb.label" ng-class="{ active: $last }">
+          <a ng-if="!$last && breadcrumb.state" title="{{breadcrumb.label}}" ui-sref="{{breadcrumb.state}}({{breadcrumb.opts}})" ui-sref-opts="{ reload: true }" ng-bind-html="breadcrumb.label | truncate:30"></a>
+          <strong ng-if="!$last">&nbsp;/&nbsp;</strong>
+          <strong ng-if="$last"><a ui-sref="posts.data({ threadId: post.thread_id, start: post.position, '#': post.id })" ng-bind-html="::post.thread_title"></a></strong>
+        </span>
       </div>
       <!-- Post Body -->
       <div class="post-body" post-processing="post.body_html" style-fix="true"></div>
