+++ modules/ept-posts/db/patrol.js
@@ -3,12 +3,14 @@ var common = require(path.normalize(__dirname + '/../common'));
 var dbc = require(path.normalize(__dirname + '/db'));
 var db = dbc.db;
 var helper = dbc.helper;
+var Boom = require('boom');
 
-module.exports = function(dbc, opts) {
+module.exports = function(request, opts) {
   opts = opts || {};
   opts.limit = opts.limit || 25;
   opts.page = opts.page || 1;
   opts.offset = (opts.page * opts.limit) - opts.limit;
+  opts.limit = opts.limit + 1;
 
   var query = `
   SELECT
@@ -72,6 +74,12 @@ module.exports = function(dbc, opts) {
     LEFT JOIN threads t ON p.thread_id = t.id
     LEFT JOIN boards b ON t.board_id = b.id
     WHERE p.id = plist.id
+    AND EXISTS (
+          SELECT 1
+          FROM boards b2
+          WHERE b2.id = t.board_id
+          AND ( b2.viewable_by IS NULL OR b2.viewable_by >= $3 )
+          AND ( SELECT EXISTS ( SELECT 1 FROM board_mapping WHERE board_id = t.board_id )))
   ) post ON true
   LEFT JOIN LATERAL (
     SELECT
@@ -86,9 +94,37 @@ module.exports = function(dbc, opts) {
   `;
 
   // get total post count for this thread
-  var params = [opts.limit, opts.offset];
+  var params = [opts.limit, opts.offset, opts.priority];
   return db.sqlQuery(query, params)
-  .map(common.formatPost)
+  .map(function(post) {
+    // Build the breadcrumbs and reply
+    return request.db.breadcrumbs.getBreadcrumbs(helper.slugify(post.thread_id), 'thread', request)
+    .then(function(breadcrumbs) {
+      post.breadcrumbs = breadcrumbs;
+      return request.server.methods.common.posts.formatPost(post);
+    });
+  })
+  .then(function(posts) {
+    // hasMoreCheck
+    var hasMorePosts = false;
+    if (posts.length > request.query.limit) {
+      hasMorePosts = true;
+      posts.pop();
+    }
+
+    return {
+      limit: request.query.limit,
+      page: request.query.page,
+      hasMorePosts: hasMorePosts,
+      posts: common.cleanPosts(posts, request.auth.credentials.id, true, request, false, true)
+    };
+  })
+  // handle page or start out of range
+  .then(function(ret) {
+    var retVal = Boom.notFound();
+    if (ret.posts.length > 0) { retVal = ret; }
+    return retVal;
+  })
   .then(helper.slugify);
 }
