+++ modules/ept-posts/routes/patrol.js
@@ -1,6 +1,5 @@
 var Joi = require('@hapi/joi');
 var path = require('path');
-var Boom = require('boom');
 var common = require(path.normalize(__dirname + '/../common'));
 
 module.exports = {
@@ -15,7 +14,7 @@ module.exports = {
       }).without('start', 'page')
     },
     pre: [
-      { method: (request) => request.server.methods.auth.posts.patrol(request.auth), assign: 'viewables' },
+      //{ method: (request) => request.server.methods.auth.posts.patrol(request.auth), assign: 'viewables' },
       { method: (request) => request.server.methods.hooks.preProcessing(request) },
       [
         { method: (request) => request.server.methods.hooks.parallelProcessing(request), assign: 'parallelProcessed' },
@@ -33,35 +32,13 @@ module.exports = {
 
 function processing(request) {
   // ready parameters
-  var userId = request.auth.credentials.id;
   var opts = {
     page: request.query.page,
-    limit: request.query.limit + 1 // check if there are more posts
+    limit: request.query.limit,
+    priority: request.server.plugins.acls.getUserPriority(request.auth)
   };
-
   // retrieve posts for this thread
-  var promise = request.db.posts.patrol(opts)
-  .then(function(posts) {
-    // hasMoreCheck
-    var hasMorePosts = false;
-    if (posts.length > request.query.limit) {
-      hasMorePosts = true;
-      posts.pop();
-    }
-
-    return {
-      limit: request.query.limit,
-      page: request.query.page,
-      hasMorePosts: hasMorePosts,
-      posts: common.cleanPosts(posts, userId, true, request, false, true)
-    };
-  })
-  // handle page or start out of range
-  .then(function(ret) {
-    var retVal = Boom.notFound();
-    if (ret.posts.length > 0) { retVal = ret; }
-    return retVal;
-  })
+  var promise = request.db.posts.patrol(request, opts)
   .error(request.errorMap.toHttpError);
 
   return promise;
